requirements.txt
fastapi
uvicorn
google-generativeai
sqlalchemy
psycopg2-binary
python-dotenv
pandas
pyodbc
gunicorn
